Move `Vote` context content into `RankedVoting`.
