+++ lib/flick_web/live/vote/vote_capture_live.ex
@@ -8,7 +8,6 @@ defmodule FlickWeb.Vote.VoteCaptureLive do
 
   alias Flick.RankedVoting
   alias Flick.RankedVoting.Ballot
-  alias Flick.Votes
   alias Flick.RankedVoting.Vote
   alias Phoenix.LiveView.Socket
 
@@ -35,20 +34,20 @@ defmodule FlickWeb.Vote.VoteCaptureLive do
     ranked_answers = Enum.map(1..ranked_answer_count, fn _ -> %{value: nil} end)
 
     vote_params = %{ballot_id: ballot.id, ranked_answers: ranked_answers}
-    changeset = Votes.change_vote(%Vote{}, vote_params)
+    changeset = RankedVoting.change_vote(%Vote{}, vote_params)
     assign(socket, form: to_form(changeset))
   end
 
   @impl Phoenix.LiveView
   def handle_event("validate", %{"vote" => vote_params}, socket) do
-    changeset = Votes.change_vote(%Vote{}, vote_params, action: :validate)
+    changeset = RankedVoting.change_vote(%Vote{}, vote_params, action: :validate)
     {:noreply, assign(socket, form: to_form(changeset))}
   end
 
   def handle_event("save", %{"vote" => vote_params}, socket) do
     %{ballot: ballot} = socket.assigns
 
-    case Votes.record_vote(ballot, vote_params) do
+    case RankedVoting.record_vote(ballot, vote_params) do
       {:ok, _vote} ->
         socket
         |> put_flash(:info, "Vote recorded.")
